requirements.txt
dj-database-url==0.4.1
Django==1.11.1
gunicorn==19.6.0
psycopg2==2.7.1
whitenoise==3.2
djangorestframework==3.4
validate-email==1.3
djangorestframework-jwt==1.11.0
django-celery==3.2.2
requests==2.18.4
pika==0.11.2
django-kombu==0.9.4
